requirements.txt
# Only use Cray-procured packages
--extra-index-url https://artifactory.algol60.net/artifactory/csm-python-modules/simple
--trusted-host artifactory.algol60.net
-c constraints.txt

ims-python-helper
requests
pyyaml
boto3
jinja2
